replaced game id with join code on frontend
