+++ apps/frontend/messages/hu.json
@@ -52,8 +52,7 @@
     "id": "Játék azonosító",
     "timer": "Időkorlát",
     "difficulty": "Nehézség",
-    "area": "Terület",
-    "share_message": "Csatlakozz hozzám játszani"
+    "area": "Terület"
   },
   "GuessMarker": {
     "my": "Tippem",
@@ -77,7 +76,11 @@
   },
   "SelectGame": {
     "create": "Játék létrehozása",
-    "id": "Játék azonosító",
+    "id": "Belépési kód",
     "join": "Belépés a játékba"
+  },
+  "JoinCode": {
+    "join_code": "Belépési kód",
+    "share_message": "Csatlakozz hozzám játszani"
   }
 }
